Added a new Variable Modifier of ‘V’ for Variance
The first character following the ‘v’ will be used as a delimiter. Everything between the delimiters will be used as a from field. Everything following the delimiter will be used as a to field. All output lines will be scanned for occurrences of the from field, and will be replaced with the to field.